canovara-robot/requirements.txt
beautifulsoup4
requests
python-telegram-bot==20.3
python-dotenv
aiohttp
brotli
